Obtain the chart group from the chartcode

Obtain the chart group from the chartcode

get_chartgrp(chartcode = NULL)

Arguments

chartcode
Chart codes, usually constructed by create_chartcode(). If not specified, the function uses all chart codes.

Value

The function returns a named vector with length(chartcode) elements with the chart group names.

Details

Warning: This function uses the property that the chartgrp can be uniquely be identified from the chartcode. When the same chartcode's are used in different chart groups, this function will give unpredictable results.

See also

Examples

get_chartgrp(c("NJAA", "NJBA"))

#>     NJAA     NJBA 
#> "nl2010" "nl2010"